Put ResultStream into its own files
